Organize component types; Breakdown main (center) column
