+++ tias/src/Components/App.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
-import { EmployeeList } from './EmployeeList';
-import { ScheduleList } from './ScheduleList';
-import { ScheduleWindow } from './ScheduleWindow';
+import { EmployeeList } from './EmployeeList/EmployeeList';
+import { ScheduleList } from './ScheduleList/ScheduleList';
+import { SchedulingWindow } from './Scheduling/SchedulingWindow';
 import './common.scss';
 
 export const App = () => {
   return (
     <div className="App">
       < EmployeeList width='25vh' />
-      < ScheduleWindow />
-      < ScheduleList width='20vh'/>
+      < SchedulingWindow />
+      < ScheduleList width='20vh' />
     </div>
   );
 }
