+++ tias/src/Components/EmployeeList/AcceptButton.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const AcceptButton = () => {
+  return (
+	  <button className="green button">Accept</button>
+  )
+}
