+++ tias/src/Components/EmployeeList/EmployeeRow.tsx
@@ -1,24 +1,26 @@
 import React, { FC } from 'react'
-import './common.scss';
 
 interface Props {
 	element: string
 }
 
-export const ScheduleRow: FC<Props> = ({element}) => {
-	return (
-		<>
+export const EmployeeRow: FC<Props> = ({element}) => {
+  return (
+    <>
       <div className="row">
         <div className="element left">
           []
         </div>
-        <div className="element right">
+        <div className="element label">
          {element}
+        </div>
+        <div className="element right">
+          •
         </div>
       </div>
       <div className="hr-container">
         <hr/>
       </div>
     </>
-	)
+  )
 }
