+++ tias/src/Components/EmployeeList/GenerateButton.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const GenerateButton = () => {
+  return (
+	  <button className="blue button">Generate</button>
+  )
+}
