+++ tias/src/Components/ScheduleList/ScheduleRow.tsx
@@ -1,27 +1,23 @@
 import React, { FC } from 'react'
-import './common.scss';
 
 interface Props {
 	element: string
 }
 
-export const EmployeeRow: FC<Props> = ({element}) => {
-  return (
-    <>
+export const ScheduleRow: FC<Props> = ({element}) => {
+	return (
+		<>
       <div className="row">
         <div className="element left">
           []
         </div>
-        <div className="element label">
-         {element}
-        </div>
         <div className="element right">
-          •
+         {element}
         </div>
       </div>
       <div className="hr-container">
         <hr/>
       </div>
     </>
-  )
+	)
 }
