+++ tias/src/Components/Scheduling/SchedulingFilter.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const SchedulingFilter = () => {
+  return (
+	  <div className="row">
+      Filter
+    </div>
+  )
+}
