+++ tias/src/Components/Scheduling/SchedulingHeader.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const SchedulingHeader = () => {
+  return (
+	  <div className="row">
+		  <h2 className="slim">Header</h2>
+    </div>
+  )
+}
