+++ tias/src/Components/Scheduling/SchedulingRender.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const SchedulingRender = () => {
+  return (
+	  <div className="render-window">
+      <div>Render</div>
+    </div>
+  )
+}
